Create destroy on collision component/systems.

+++ integrations/libgdx/src/main/kotlin/dev/andrewhan/nomo/integration/libgdx/render/components/CameraComponent.kt
@@ -40,8 +40,8 @@ class CameraComponent(
       viewport.setScreenBounds(
         viewportOrigin.x,
         viewportOrigin.y,
-        viewportSize.width,
-        viewportSize.height
+        viewportSize.width.toInt(),
+        viewportSize.height.toInt()
       )
       viewport.apply()
       block(viewport.camera)
